Move helper functions from LPI to analysi.base
